[#46] refactor : 객체명 변경
SBO 보다는 Inning에 대한 status라고 표현해주는 것이 낫겠음.
